Runtime permissions request (for Android 6 only)
See https://github.com/agamemnus/cordova-plugin-xapkreader/issues/94

In Android 6 (and only Android 6) we need WRITE_EXTERNAL_STORAGE
permission in order to read the OBB file. This is a known Android
6 bug; the OBB file is supposed to be part of the app's internal
storage. See https://code.google.com/p/android/issues/detail?id=197287

In Android target api >= 23, WRITE_EXTERNAL_STORAGE is considered
part of the "dangerous" STORAGE permissions group. So, instead of
asking the user to grant it when the app is installed/upgraded, we
have to prompt the user for it at runtime. That's what this patch
adds.